source/requirements.txt
alabaster==0.7.12
appdirs==1.4.3
argh==0.26.2
asgiref==3.3.1
astroid==2.3.3
atomicwrites==1.3.0
attrs==19.3.0
autopep8==1.4.4
Babel==2.8.0
backcall==0.1.0
bcrypt==3.1.7
beautifulsoup4==4.8.1
bleach==3.1.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cloudpickle==1.2.2
colorama==0.4.1
cryptography==2.8
decorator==4.4.1
defusedxml==0.6.0
diff-match-patch==20181111
distlib==0.3.0
dj-database-url==0.5.0
Django==3.1.5
django-heroku==0.3.1
django-tinymce==3.2.0
dnspython==1.16.0
docutils==0.16
entrypoints==0.3
filelock==3.0.12
flake8==3.7.9
future==0.18.2
gunicorn==20.0.4
helpdev==0.6.10
idna==2.8
imagesize==1.2.0
importlib-metadata==0.23
intervaltree==3.0.2
ipykernel==5.1.3
ipython==7.9.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
jedi==0.14.1
Jinja2==2.10.3
jsonschema==3.1.1
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.1
keyring==21.1.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
more-itertools==7.2.0
mysql-connector-python==8.0.19
nbconvert==5.6.1
nbformat==4.4.0
notebook==6.0.1
numpydoc==0.9.2
packaging==19.2
pandocfilters==1.4.2
paramiko==2.7.1
parso==0.5.1
pathtools==0.1.2
pexpect==4.7.0
pickleshare==0.7.5
Pillow==7.0.0
pluggy==0.13.1
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.6.1
psutil==5.6.7
psycopg2==2.8.6
ptyprocess==0.6.0
py==1.8.0
pycodestyle==2.5.0
pycparser==2.19
pydocstyle==5.0.2
pyflakes==2.1.1
Pygments==2.4.2
pylint==2.4.4
PyNaCl==1.3.0
pyparsing==2.4.5
PyQt5==5.12.3
PyQt5-sip==12.7.0
PyQtWebEngine==5.12.1
pyrsistent==0.15.5
pytest==5.3.1
python-dateutil==2.8.1
python-decouple==3.4
python-jsonrpc-server==0.3.2
python-language-server==0.31.4
pytz==2019.3
pywin32==225
pywin32-ctypes==0.2.0
pywinpty==0.5.5
PyYAML==5.3
pyzmq==18.1.0
QDarkStyle==2.7
QtAwesome==0.6.1
qtconsole==4.6.0
QtPy==1.9.0
requests==2.22.0
rope==0.16.0
Send2Trash==1.5.0
six==1.12.0
snowballstemmer==2.0.0
sortedcontainers==2.1.0
soupsieve==1.9.5
Sphinx==2.3.1
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
spyder==4.0.1
spyder-kernels==1.8.1
sqlparse==0.4.1
terminado==0.8.2
testpath==0.4.3
tornado==6.0.3
traitlets==4.3.3
typed-ast==1.4.1
urllib3==1.25.7
virtualenv==20.0.0
watchdog==0.9.0
wcwidth==0.1.7
webencodings==0.5.1
whitenoise==5.2.0
widgetsnbextension==3.5.1
wrapt==1.11.2
xmltodict==0.12.0
yapf==0.29.0
zipp==0.6.0
